renames offer-card to place-card
